+++ java/client/src/main/java/glide/api/RedisClient.java
@@ -18,6 +18,7 @@ import static redis_request.RedisRequestOuterClass.RequestType.Info;
 import static redis_request.RedisRequestOuterClass.RequestType.LastSave;
 import static redis_request.RedisRequestOuterClass.RequestType.Lolwut;
 import static redis_request.RedisRequestOuterClass.RequestType.Ping;
+import static redis_request.RedisRequestOuterClass.RequestType.RandomKey;
 import static redis_request.RedisRequestOuterClass.RequestType.Select;
 import static redis_request.RedisRequestOuterClass.RequestType.Time;
 
@@ -202,5 +203,11 @@ public class RedisClient extends BaseClient
                 FunctionLoad,
                 new String[] {FunctionLoadOptions.REPLACE.toString(), libraryCode},
                 this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> randomKey() {
+        return commandManager.submitNewCommand(
+            RandomKey, new String[0], this::handleStringOrNullResponse);
     }
 }
